requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
sqlalchemy
psycopg2
databases[aiosqlite]
databases[asyncpg]
python-dotenv
python-json-logger
python-jose
python-multipart
passlib[bcrypt]
apscheduler
